get posts and comments from db
