--- backend/src/api/posts/getPostById.ts
@@ -1,3 +1,0 @@
-export default async function getPostById(req, res) {
-    res.send('getPostById');
-}
