+++ backend/src/app.ts
@@ -2,7 +2,7 @@ import helmet from "helmet";
 import express, { Application, Request, Response } from "express"
 import categorize from "./api/categorize/categorize"
 import cors from 'cors';
-import postsRouter from './api/posts/router';
+import posts from './api/posts/posts';
 import FacebookClient from "./utils/facebookClient";
 import prisma from "./utils/prisma";
 import jwt from "jsonwebtoken"
@@ -110,7 +110,7 @@ app.get("/me", authMiddleware(), async (req: Request, res, next) => {
 })
 
 app.use("/categorize", categorize);
-app.use("/posts", postsRouter);
+app.use("/posts", posts);
 
 const server = app.listen(PORT, () => {
   console.log(`Environment: ${process.env.NODE_ENV}`);
